% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{addEqualSign}
\alias{addEqualSign}
\title{Add equal sign to begining of paramenter}
\usage{
addEqualSign(x)
}
\arguments{
\item{x}{string}
}
\value{
string
}
\description{
Add equal sign to begining of paramenter
}
\examples{
addEqualSign('EP')
}
