% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{readFile}
\alias{readFile}
\title{Read a text file}
\usage{
readFile(filePath = "")
}
\arguments{
\item{filePath}{path to file to read}
}
\value{
string representation of text file
}
\description{
Read a text file
}
\examples{
readFile(filePath = 'example.sql')
}
