% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{singleQuote}
\alias{singleQuote}
\title{Add single quotes around x}
\usage{
singleQuote(x)
}
\arguments{
\item{x}{string}
}
\value{
string
}
\description{
Add single quotes around x
}
\examples{
singleQuote("EP")
}
