% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sqlFileExecute}
\alias{sqlFileExecute}
\title{Return the results of running SQL file}
\usage{
sqlFileExecute(sqlFile, stringsAsFactors = TRUE, connectVar)
}
\arguments{
\item{sqlFile}{String. The path to your .sql file}

\item{stringsAsFactors}{Boolean. Will strings be converted to factors?}

\item{connectVar}{RODBC Connection. Name of the object holding your ODBC connection.}
}
\value{
Data frame containing the results of your SQL query executed agaist
the named channel.
}
\description{
\code{sqlFileExecute} returns the results of an SQL query, as a data frame
}
\details{
This function makes it convenient to develop your SQL in whatever IDE you
choose, without copy/pasting code to your R script--which would make
maintenance tedious.
}
\examples{
df = sqlFileExecute(sqlFile = 'file/path/exampleQuery.sql', stringsAsFactors = FALSE)
}
