% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{whereIn}
\alias{whereIn}
\title{Generate a where in clause}
\usage{
whereIn(inputVals, filterCol, dropIfIncludes = NULL)
}
\arguments{
\item{inputVals}{list of values to put after "WHERE IN"}

\item{filterCol}{SQL database column name to put before "WHERE IN"}

\item{dropIfIncludes}{Optional value}
}
\value{
string
}
\description{
Generate a where in clause
}
\examples{
whereIn(input$divFilter)
}
